playground/archive/online-learning-game/training/requirements.txt
google-cloud-aiplatform>=1.70.0
google-cloud-storage>=2.18.0
python-dotenv>=1.0.0
